Last Commit

+++ src/main/java/com/enoca/challenge/entities/Student.java
@@ -1,5 +1,6 @@
 package com.enoca.challenge.entities;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.Data;
 
@@ -12,7 +13,7 @@ public class Student {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private int id;
 
     @Column(name = "firstname")
     private String firstname;
