+++ src/main/java/com/enoca/challenge/repositories/UniversityRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UniversityRepo extends JpaRepository<University,Long> {
+public interface UniversityRepo extends JpaRepository<University,Integer> {
 }
