+++ src/main/java/com/enoca/challenge/services/UniversityService.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 public interface UniversityService {
 
     public University save(University university);
-    public Optional<University> getOneById(Long id);
+    public Optional<University> getOneById(int id);
     public List<University> getAll();
-    public void delete(Long id);
+    public void delete(int id);
 }
